Refactor Food controller and service logic
The services were responsible for throwing HTTP error codes, which was not a
good sign.

Change the logic, services now throw InvalidDataException which is
translated to HTTP code at controller level, thus decoupling the logic.

Add transactional annotation when saving food. When exception would be thrown,
foodTags will be rollbacked so it prevents the state when new tags would be
saved, but food containing these tags would fail to save.